добавит getData и вынесет в него получение даных из pokedex
